fix(R.5): remove retained runtime shim surface
